+++ assets/js/tokenize.js
@@ -1,7 +1,5 @@
 jQuery(document).ready(function($) {
-	// Conekta.setPublishableKey(wc_conekta_params.public_key);
-	var current_account_token = $("#current_account_token").val()
-	Conekta.setPublishableKey(current_account_token);
+	Conekta.setPublishableKey(wc_conekta_params.public_key);
 
 	var $form = $('form.checkout,form#order_review');