+++ conekta_spei_gateway.php
@@ -28,7 +28,6 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
     public $account_owner;
     public $secret_key;
     public $lang_options;
-    public $account = false;
 
     public function __construct()
     {
@@ -83,21 +82,36 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
         header('HTTP/1.1 200 OK');
         $body          = @file_get_contents('php://input');
         $event         = json_decode($body, true);
+        
+         // Respondiendo a eventos "ping"
+        if (isset($event['type']) && $event['type'] === 'webhook_ping') {
+            header('Content-Type: application/json');
+            echo json_encode(['message' => 'OK']);
+            exit; // Salir de la función después de manejar el evento ping
+        }
+
         $conekta_order = $event['data']['object'];
         $charge        = $conekta_order['charges']['data'][0];
         $order_id      = $conekta_order['metadata']['reference_id'];
-        $paid_at       = date("Y-m-d", $charge['paid_at']);
         $order         = new WC_Order($order_id);
 
-        if (strpos($event['type'], "order.paid") !== false
+       
+        if ( $event['type']=== "order.paid" 
             && $charge['payment_method']['type'] === "spei")
-            {
-                update_post_meta( $order->get_id(), 'conekta-paid-at', $paid_at);
-                $order->payment_complete();
-                $order->add_order_note(sprintf("Payment completed in Spei and notification of payment received"));
+        {
+            $paid_at = date("Y-m-d", $charge['paid_at']);
+            update_post_meta( $order->get_id(), 'conekta-paid-at', $paid_at);
+            $order->payment_complete();
+            $order->add_order_note(sprintf("Payment completed in Spei and notification of payment received"));
 
-                parent::ckpg_offline_payment_notification($order_id, $conekta_order['customer_info']['name']);
-            }
+            parent::ckpg_offline_payment_notification($order_id, $conekta_order['customer_info']['name']);
+        }
+         // expired orders
+         if ( ($event['type'] === "order.expired"  || $event['type'] ==="order.canceled")
+            && $charge['payment_method']['type'] === "spei") 
+        {
+            $order->update_status('cancelled', 'Order expired in Conekta.');
+        }
     }
 
     public function ckpg_init_form_fields()
@@ -178,15 +192,14 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
      * @param WC_Order $order
      */
     function ckpg_email_reference($order) {
-	    // Comentado por danielgc para ponerlo directamente en el template de Woo
-        
-        // if (get_post_meta( $order->get_id(), 'conekta-clabe', true ) != null)
-        //     {
-        //         echo '<p><h4><strong>'.esc_html(__('Clabe')).':</strong> '
-        //         . esc_html( get_post_meta( $order->get_id(), 'conekta-clabe', true ) ). '</h4></p>';
-        //         echo '<p><h4><strong>'.esc_html(__('Beneficiario')).':</strong> '.esc_html($this->account_owner).'</h4></p>';
-        //         echo '<p><h4><strong>'.esc_html(__('Banco Receptor')).':</strong>  Sistema de Transferencias y Pagos (STP)<h4></p>';
-        //     }
+
+        if (get_post_meta( $order->get_id(), 'conekta-clabe', true ) != null)
+            {
+                echo '<p><h4><strong>'.esc_html(__('Clabe')).':</strong> '
+                . esc_html( get_post_meta( $order->get_id(), 'conekta-clabe', true ) ). '</h4></p>';
+                echo '<p><h4><strong>'.esc_html(__('Beneficiario')).':</strong> '.esc_html($this->account_owner).'</h4></p>';
+                echo '<p><h4><strong>'.esc_html(__('Banco Receptor')).':</strong>  Sistema de Transferencias y Pagos (STP)<h4></p>';
+            }
     }
 
     /**
@@ -216,8 +229,6 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
 
     protected function ckpg_send_to_conekta()
     {
-        mg_gateways_set_current_bank_account( $this->order, $this );
-
         global $woocommerce;
         include_once('conekta_gateway_helper.php');
         \Conekta\Conekta::setApiKey($this->secret_key);
@@ -236,7 +247,11 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
         $shipping_contact = ckpg_build_shipping_contact($data);
         $tax_lines        = ckpg_build_tax_lines($taxes);
         $customer_info    = ckpg_build_customer_info($data);
-        $order_metadata   = ckpg_build_order_metadata($data);
+        $order_metadata   = ckpg_build_order_metadata($data + array(
+                                                                    'plugin_conekta_version' => $this->version,
+                                                                    'woocommerce_version'   => $woocommerce->version,
+                                                                    )
+                                                    );
         $order_details    = array(
             'currency'         => $data['currency'],
             'line_items'       => $line_items,
@@ -266,9 +281,6 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
             }
 
             update_post_meta($this->order->get_id(), 'conekta-order-id', $order->id);
-            update_post_meta( $this->order->get_id(), 'additional_branch_track', mg_format_additional_branch_track( $this->account, $this->order->get_id(), $order->id ) );
-
-            $this->order->add_order_note( 'Realizando pago para: ' . $this->account['name'] );
 
             $charge_details = array(
                 'payment_method' => array('type' => 'spei'),
@@ -282,9 +294,6 @@ class WC_Conekta_Spei_Gateway extends WC_Conekta_Plugin
             update_post_meta( $this->order->get_id(), 'conekta-creado', $charge->created_at );
             update_post_meta( $this->order->get_id(), 'conekta-expira', $charge->payment_method->expires_at );
             update_post_meta( $this->order->get_id(), 'conekta-clabe', $charge->payment_method->clabe );
-
-            mg_gateways_send_mail_notification( $this->order, true );
-
             return true;
         } catch(\Conekta\Handler $e) {
             $description = $e->getMessage();