+++ templates/cash_admin.php
@@ -8,10 +8,10 @@
 ?>
 
 <h3>
-    <?php _e('Oxxo Pay Payment', 'woothemes'); ?>
+    <?php _e('Cash Pay Payment', 'woothemes'); ?>
 </h3>
 
-<p><?php _e('Allows Oxxo Pay payments.', 'woothemes'); ?></p>
+<p><?php _e('Allows Cash Pay payments.', 'woothemes'); ?></p>
 
 <table class="form-table">
     <?php $this->generate_settings_html(); ?>